refactor: notification handlers
