+++ src/renderer/components/notifications/NotificationRow.tsx
@@ -9,11 +9,9 @@ import type { Notification } from '../../typesGitHub';
 import { cn } from '../../utils/cn';
 import { isMarkAsDoneFeatureSupported } from '../../utils/features';
 import { formatForDisplay } from '../../utils/helpers';
-import {
-  getNotificationTypeIcon,
-  getNotificationTypeIconColor,
-} from '../../utils/icons';
+import { getNotificationTypeIconColor } from '../../utils/icons';
 import { openNotification } from '../../utils/links';
+import { createNotificationHandler } from '../../utils/notifications/handlers';
 import { HoverButton } from '../primitives/HoverButton';
 import { HoverGroup } from '../primitives/HoverGroup';
 import { NotificationFooter } from './NotificationFooter';
@@ -73,7 +71,9 @@ export const NotificationRow: FC<INotificationRow> = ({
     unsubscribeNotification(notification);
   };
 
-  const NotificationIcon = getNotificationTypeIcon(notification.subject);
+  const handler = createNotificationHandler(notification);
+
+  const NotificationIcon = handler.getIcon(notification.subject);
   const iconColor = getNotificationTypeIconColor(notification.subject);
 
   const notificationType = formatForDisplay([
