+++ src/renderer/utils/notifications/handlers/pullRequest.ts
@@ -1,0 +1,175 @@
+import type { FC } from 'react';
+
+import type { OcticonProps } from '@primer/octicons-react';
+import {
+  GitMergeIcon,
+  GitPullRequestClosedIcon,
+  GitPullRequestDraftIcon,
+  GitPullRequestIcon,
+} from '@primer/octicons-react';
+
+import type { Link, SettingsState } from '../../../types';
+import type {
+  GitifyPullRequestReview,
+  GitifySubject,
+  Notification,
+  PullRequest,
+  PullRequestReview,
+  PullRequestStateType,
+  Subject,
+  User,
+} from '../../../typesGitHub';
+import {
+  getIssueOrPullRequestComment,
+  getPullRequest,
+  getPullRequestReviews,
+} from '../../api/client';
+import { isStateFilteredOut, isUserFilteredOut } from '../filters/filter';
+import type { NotificationTypeHandler } from './types';
+import { getSubjectUser } from './utils';
+
+class PullRequestHandler implements NotificationTypeHandler {
+  readonly type = 'PullRequest' as const;
+
+  async enrich(
+    notification: Notification,
+    settings: SettingsState,
+  ): Promise<GitifySubject> {
+    const pr = (
+      await getPullRequest(notification.subject.url, notification.account.token)
+    ).data;
+
+    let prState: PullRequestStateType = pr.state;
+    if (pr.merged) {
+      prState = 'merged';
+    } else if (pr.draft) {
+      prState = 'draft';
+    }
+
+    // Return early if this notification would be hidden by state filters
+    if (isStateFilteredOut(prState, settings)) {
+      return null;
+    }
+
+    let prCommentUser: User;
+    if (
+      notification.subject.latest_comment_url &&
+      notification.subject.latest_comment_url !== notification.subject.url
+    ) {
+      const prComment = (
+        await getIssueOrPullRequestComment(
+          notification.subject.latest_comment_url,
+          notification.account.token,
+        )
+      ).data;
+      prCommentUser = prComment.user;
+    }
+
+    const prUser = getSubjectUser([prCommentUser, pr.user]);
+
+    // Return early if this notification would be hidden by user filters
+    if (isUserFilteredOut(prUser, settings)) {
+      return null;
+    }
+
+    const reviews = await getLatestReviewForReviewers(notification);
+    const linkedIssues = parseLinkedIssuesFromPr(pr);
+
+    return {
+      number: pr.number,
+      state: prState,
+      user: prUser,
+      reviews: reviews,
+      comments: pr.comments,
+      labels: pr.labels?.map((label) => label.name) ?? [],
+      linkedIssues: linkedIssues,
+      milestone: pr.milestone,
+    };
+  }
+
+  getIcon(subject: Subject): FC<OcticonProps> | null {
+    switch (subject.state) {
+      case 'draft':
+        return GitPullRequestDraftIcon;
+      case 'closed':
+        return GitPullRequestClosedIcon;
+      case 'merged':
+        return GitMergeIcon;
+      default:
+        return GitPullRequestIcon;
+    }
+  }
+}
+
+export const pullRequestHandler = new PullRequestHandler();
+
+export async function getLatestReviewForReviewers(
+  notification: Notification,
+): Promise<GitifyPullRequestReview[]> | null {
+  if (notification.subject.type !== 'PullRequest') {
+    return null;
+  }
+
+  const prReviews = await getPullRequestReviews(
+    `${notification.subject.url}/reviews` as Link,
+    notification.account.token,
+  );
+
+  if (!prReviews.data.length) {
+    return null;
+  }
+
+  // Find the most recent review for each reviewer
+  const latestReviews: PullRequestReview[] = [];
+  const sortedReviews = prReviews.data.reverse();
+  for (const prReview of sortedReviews) {
+    const reviewerFound = latestReviews.find(
+      (review) => review.user.login === prReview.user.login,
+    );
+
+    if (!reviewerFound) {
+      latestReviews.push(prReview);
+    }
+  }
+
+  // Group by the review state
+  const reviewers: GitifyPullRequestReview[] = [];
+  for (const prReview of latestReviews) {
+    const reviewerFound = reviewers.find(
+      (review) => review.state === prReview.state,
+    );
+
+    if (!reviewerFound) {
+      reviewers.push({
+        state: prReview.state,
+        users: [prReview.user.login],
+      });
+    } else {
+      reviewerFound.users.push(prReview.user.login);
+    }
+  }
+
+  // Sort reviews by state for consistent order when rendering
+  return reviewers.sort((a, b) => {
+    return a.state.localeCompare(b.state);
+  });
+}
+
+export function parseLinkedIssuesFromPr(pr: PullRequest): string[] {
+  const linkedIssues: string[] = [];
+
+  if (!pr.body || pr.user.type !== 'User') {
+    return linkedIssues;
+  }
+
+  const regexPattern = /\s?#(\d+)\s?/gi;
+  const matches = pr.body.matchAll(regexPattern);
+
+  for (const match of matches) {
+    if (match[0]) {
+      linkedIssues.push(match[0].trim());
+    }
+  }
+
+  return linkedIssues;
+}
