+++ src/renderer/utils/notifications/handlers/repositoryVulnerabilityAlert.ts
@@ -1,0 +1,29 @@
+import type { FC } from 'react';
+
+import { AlertIcon, type OcticonProps } from '@primer/octicons-react';
+
+import type { SettingsState } from '../../../types';
+import type {
+  GitifySubject,
+  Notification,
+  Subject,
+} from '../../../typesGitHub';
+import type { NotificationTypeHandler } from './types';
+
+class RepositoryVulnerabilityAlertHandler implements NotificationTypeHandler {
+  readonly type = 'RepositoryVulnerabilityAlert';
+
+  async enrich(
+    _notification: Notification,
+    _settings: SettingsState,
+  ): Promise<GitifySubject> {
+    return;
+  }
+
+  getIcon(_subject: Subject): FC<OcticonProps> | null {
+    return AlertIcon;
+  }
+}
+
+export const repositoryVulnerabilityAlertHandler =
+  new RepositoryVulnerabilityAlertHandler();
